Refactor features to allow any scalar to have a mapping. Also splitted some code away from scalars. Add public Features to __init__.py.
